Fix handling of paused status in Pingdom
This handles the discrepancy between Pingdom read/write API for checks
(`paused` is a bool parameter on write, but indicated by specific value
of `status` field on read) by using the `status` field when comparing
_spec_ to _status_.

Also add constants for allowed values of `status` and remove the
`paused` field from `CheckStatus` as it's not necessary there and could
be confusing.

Finally add a desired `paused` setting to the sample spec so that its
reconciliation can be seen in action with the default sample.