Fix handling of empty UserIds on read
When UserIds are not set go-pingdom returns an uninitialised int slice
on read which we were then using directly in check status object.
However that fails validation when trying to update the status
subresource via the API.

Change read method to always return an initialised but possibly empty
slice for UserIds.